Extract DB command repository
